TextureAPI/requirements.txt
GitPython

# Podstawowe ML
diffusers[torch]==0.26.3
transformers==4.30.2
accelerate==0.21.0
safetensors==0.4.1
huggingface_hub==0.20.2
git+https://github.com/cloneofsimo/lora.git@master#egg=lora_diffusion

# GUI / GTK
PyGObject==3.44.1
pycairo==1.25.0
Pillow
pillow-avif-plugin==1.4.3
scikit-image>=0.19

# API
fastapi==0.110.0
uvicorn[standard]==0.29.0
einops
jsonmerge
numpy
requests

